data/data_ingestion_requirements.txt
requests
beautifulsoup4
pandas
numpy
pyyaml
pytest